Re-introduce Makefile support for hier blocks
The hierarchical blocks are now part of this (gr-blockstream)
repository. Move their build routine from the original repository into
the Makefile of this project and update the paths to comply with their
new location (at apps/hier).

Additionally, update the hierarchical grc sources to maximize the
similarity of what grcc generates and what is currently committed as
.xml and .py (grcc results) in the project tree. Moreover, update some
parts of the pre-generated python files in the "python/" folder
according to how grcc generates them. The grcc results have changed
slightly once "mods" was renamed to "blockstream" due to alphabetical
line sorting.