Add demultiplexer for user and blocks data
Substitute the asynchronous FIFO hierarchical block by a single (basic) block
that plays the role of a data demultiplexer for the Blockstream Satellite
protocol. That is, add the "protocol sink" block, which is responsible for
separating user and blocks traffic.

The previous hierarchical async FIFO block had internally a block to convert
from PDU to tagged stream. Now, with the "protocol sink" block, instead of
relying on conversion from PDU to tagged stream, the block receives PDUs
directly. This ultimately allows reception of variable length PDUs, as the PDU
size does not need to be fixed by the PDU-to-tagged-stream converter.

To accomplish the block substitution, update also all corresponding CMake lists
and the parameters in the GRC XML.