Output data received via API within a structure
Support the "more fragments" bit introduced in the first octet of the Blocksat
Packet header and, using this indication, accumulate the data from all related
fragments until the last fragment is received. Once all fragments are
accumulated, place the entire data buffer into a pre-defined data structure and
output to named pipe. This allows applications to read from API pipe and
successfully detect their individual transmissions sent via API.

The added output data structure is composed of a header and the message
itself (the payload). The header, in turn, consists of a delimiter (i.e. a magic
sequence) that applications can search and recognize. It also includes the
message length, so that once an application finds the delimiter, it can also
know how much data it should read ahead from that point.