Fix central frequency skipped in scan mode
The central frequency set by the user via command-line argument was being tested
only in the first run of the scan mode frequency sweep. It was being skipped in
subsequent runs due to the fact that right after resetting the frequency to the
central one, the frequency was changed to the "next value" before testing the
central one first. This patch fixes this problem.