Add mechanism in Rx logger to update GUI labels
Convey setters for variables that are used in GUI labels directly to the Rx
logger instance and use these setters to update the GUI label values. The
updates are carried in the same print rate that the Rx logger uses for the given
metric of interest.

Add support for logging the BER and the Frame Sync state into GUI labels.