Add optimized QPSK/BPSK soft decoder block
Add a block that computes the analytical (floating-point precision) LLRs for the
QPSK and BPSK constellations that are used in the Blocksat receiver. The
computations are optimized for the specific constellations of interest. This is
preferable to using the generic soft decoder, which computes distances with
respect to all constellation points for each incoming symbol (CPU-intense
approach), or else needs to rely on fixed-point precision LLR look-up
tables (performance-limited approach).