Add automatic gain control (AGC) block
Add AGC block based on the AGC implementation from the standard GNU Radio
library. Customize the error detector such that the error becomes the difference
between the reference variance and the magnitude squared of the incoming
symbol. This differs from the GR block, where the error is the difference
between a reference value and the magnitude (not squared) of the incoming
symbol. By using the customized error detector, the gain of the loop converges
to the mean square of the symbols, namely the variance of the symbols, given
they are zero-mean. With that, the square root of the gain becomes the RMS
value. Ultimately, by using the square root of the gain to scale the symbols,
the reference value that the loop approaches is indeed the reference RMS value,
as desired in the first place.

Unlike the GR block, this AGC block also has an optional output stream port for
the signal level. The latter is used in plots of the GUI flowgraphs.