Improve frame sync PMF peak phase resolution
The previous approach looked for a high number in one of real or imaginary
dimension and a low enough number on the other dimension, given the peak is
expected to be purely real or purely imaginary (positive or negative). However,
the "low enough" number on dimension other than the dimension where the peak
appears was inferred based on threshold. For low SNR, it seems that the adopted
threshold was not enough.

Now, instead of doing so, we rotate the PMF peak complex value such that it
becomes positioned in the complex plane similarly to a QPSK constellation. We
then apply the QPSK slicing strategy and resolve the phase correction value from
there.