Support 1/2 Turbo code rate using (de)puncturer
Add support for using a Turbo depuncturer in correspondence to a Turbo puncturer
on the Tx side such that the 1/3 code rate is altered to 1/2 code rate. Allow
the Turbo code rate to be configurable at the top level by adding options to the
XML block description. Also, remove the parameter that assigns the codeword
length, given the codeword length is not freely adjustable. Instead, define the
codeword internally based on the dataword length and the target code rate.