Revert ML phase error detector on data-aided mode
Revert the maximum likelihood (ML) phase error detector that was introduced for
all frame sections. Use the ML detector only on the data symbols. In contrast,
on the data-aided sections (preamble and tracking symbols), use the atan
detector. The motivation is to avoid the ambiguity at error pi rad that affects
the ML detector.