Ensure coarse CFO correction only on frame start
Introduce message passing mechanism between the frame synchronizer and the CFO
recovery block. The frame synchronizer uses this message channel to inform the
CFO recovery block about the frame start index. This information is provided in
terms of symbols. Since the CFO recovery block operates on samples, it maps the
informed symbol index to a corresponding sample index and saves this information
internally.

Using the notion of the sample where the frame starts, the coarse frequency
recovery block tries to update the frequency correction value only on that
particular sample index. This ensures that abrupt frequency changes are applied
in the preamble only, so that there is the entire preamble for the data-aided
phase recovery to adapt to the new frequency.

The process in which the CFO recovery block learns the frame start sample index
is iterative. First, the frame synchronizer sends the symbol index corresponding
to the frame start that it locked to. As mentioned earlier, the CFO recovery
maps this to a sample index and uses that index. Whenever it updates the CFO
correction value, it indicates downstream the sample where it applied the
update. The frame synchronizer, in turn, checks the incoming tagged symbol and
assesses whether it is indeed close to the frame start symbol index. Due to the
resampler stage in between the two blocks, which includes filtering and group
delay, the index won't be accurate in the beginning. However, the frame
synchronizer detects the misalignment and reports a new start index for the CFO
recovery block. After a few iterations, the index used by the CFO recovery block
becomes more accurate, according to what can be seen by the frame synchronizer.