Add data-aided fine frequency offset estimation
Estimate the frequency offset at a fine level by relying on the known symbols
that compose the preamble. Do so at the frame synchronizer and report the
estimation downstream via stream tag. Use the tag at the DA carrier phase
recovery block in order to initialize the integrator of the phase recovery loop
with a better initial guess on every frame. Use also the fine frequency offset
estimation in the frame synchronizer itself in order to de-rotate the preamble
symbols prior to the cross-correlation computation (PMF filtering). This should
alleviate frequency offset effects on the magnitude of the PMF peak.

The fine offset estimation and reporting functionality can be optionally
disabled via block parameter. Furthermore, it can be debugged via the frame sync
verbosity parameter.