Improve debug statistics of phase recovery block
Allow more fine grained debugging of SNR along the frame by introducing SNR
measurement reports for each individual data-aided segment. Report the SNR of
the preamble, the SNR of each tracking segment individually, the average SNR of
all tracking segments together and the average SNR including preamble and all
tracking segments. This should reveal how the SNR varies over time along the
frame and the impact of this variation on the all-time average SNR.

Improve also the format of the debug logs to facilitate inspection.