Support disabling of the MER measurement block
Add mechanism to disable the computations taken at the MER measurement
block. When this block operates on console-only flowgraphs, the measurements
that it takes are only displayed during initialization. After frame lock, the
SNR measurements are instead read directly from the data-aided measurements that
are taken at the carrier phase recovery block. Hence, the MER measurement
becomes useless after frame lock. In contrast, on the GUI flowgraphs, the
measurements are displayed in the GUI.

The Rx logger can disable the MER measurement block when frame lock is
achieved. To ensure that this disabling only affects the console-only flowgraph,
the presence of an output is detected before considering the disabling. It is
assumed that the console-only only flowgraphs won't have the output port (now
made optional) connected to anything, whereas the GUI flowgraphs will.

If frame sync is lost, the MER measurement block is accordingly re-enabled by
the Rx logger, who would have previously disabled it.