Change BER prints to HDLC frame error rate
Instead of printing the bit error rate (BER) computed by the HDLC deframer,
print the frame error rate (FER), which is the same as the frame drop rate. The
motivation is that the way that the BER is computed is not accurate and can be
misleading. Every time an HDLC frame is dropped due to CRC error, the number of
bit errors increases by the total number of bits in the dropped HDLC frame. In
reality, however, the CRC could have failed due to a single bit error. In
contrast, the FER is an accurate measurement. By logging both the FER and the
actual number of dropped frames, it becomes easier to debug by inspection when
exactly the frames are being dropped.