Restrict frame start index to acceptable range
There is a problematic range to lock frame timing in the frame synchronizer,
which is when the frame start index lies within the last "preamble_len - 1"
indexes of the frame window. More specifically, the problem comes when the index
is such that the preamble won't be entirely found in the frame window after the
start index. This patch ensures the block does not lock within this
range. Instead, the block re-aligns the window and puts the preamble on the
start of the window. Care is taken to ensure that the information of frame start
sent by the frame synchronizer upstream (to CFO recovery block) is still valid
after such a re-alignment.

that the frame synchronizer searches for a tag from the CFO recovery block. It
is safer to search over the entire frame window. Besides, this does not
introduce significant CPU usage penalty.