Fix initialization of CFO recovery rotator phasor
Volk's rotator used to shift frequencies and correct the carrier frequency
offset (CFO) relies on a phasor that represents the phase compensation per
sample. This phasor is a complex exponential with unitary magnitude. However, it
was previously wrongly initialized as 0.

This was not a problem most of the times, because the CFO recovery block
normally estimates a CFO right in the beginning and, as a result, overwrites the
compensation phasor before using it for the first time. Still, it would be a
problem on a setup that finds 0 CFO (very unlikely) or, more likely, on a setup
that is still not pointed correctly and thus doesn't find any signal to estimate
CFO on.